requirements.txt
fastapi
uvicorn
motor
pydantic[email]
python-dotenv
langchain
langchain-community
langchain-core
langchain-google-genai
google-generativeai
langchain-huggingface
transformers
